clippy: Apply Rust 1.94 clippy fixes across workspace #BDK-367
- collapse nested `if let`/`if` blocks into let-chains with `&&`
- replace `io::Error::new(ErrorKind::Other, ...)` with `io::Error::other(...)`
- derive `Default` with `#[default]` attribute instead of manual impls
- use `Ipv4Addr::UNSPECIFIED` instead of `Ipv4Addr::new(0, 0, 0, 0)`
- use `f32::midpoint()` instead of `(a + b) / 2.0`
- remove unnecessary `#[must_use]` on `LedDriverFactory::new()`
- remove extraneous parentheses in `Pin<Box<dyn ...>>` type
- use `checked_sub().unwrap()` for `Duration` subtraction in test
- use `.is_multiple_of(60)` instead of `% 60 == 0`

+++ bmc-upgrade/src/downloader.rs
@@ -115,10 +115,10 @@ impl DownloadWriter<BufWriter<Cursor<Vec<u8>>>> {
 
 impl<T: AsyncWrite + Unpin + Send + Sync + 'static> Drop for DownloadWriter<T> {
     fn drop(&mut self) {
-        if !self.download_finished {
-            if let Some(ref path) = self.path {
-                let _ = std::fs::remove_file(path);
-            }
+        if !self.download_finished
+            && let Some(ref path) = self.path
+        {
+            let _ = std::fs::remove_file(path);
         }
     }
 }
